refactor: Update package structure according to the domain set
