+++ s3-source-connector/src/main/java/sourceconnector/domain/batch/MessageBatch.java
@@ -1,4 +1,4 @@
-package sourceconnector.domain;
+package sourceconnector.domain.batch;
 
 import java.util.List;
