+++ s3-source-connector/src/main/java/sourceconnector/domain/factory/OffsetRecordFactory.java
@@ -1,7 +1,7 @@
 package sourceconnector.domain.factory;
 
-import sourceconnector.domain.MessageBatch;
-import sourceconnector.domain.OffsetRecord;
+import sourceconnector.domain.batch.MessageBatch;
+import sourceconnector.domain.offset.OffsetRecord;
 import sourceconnector.domain.log.Log;
 
 public interface OffsetRecordFactory<T extends Log> {
