+++ s3-source-connector/src/main/java/sourceconnector/domain/offset/FileOffsetRecord.java
@@ -1,4 +1,4 @@
-package sourceconnector.domain;
+package sourceconnector.domain.offset;
 
 public record FileOffsetRecord(
   String key,
