+++ s3-source-connector/src/main/java/sourceconnector/domain/offset/LocalFileOffsetRecord.java
@@ -1,4 +1,4 @@
-package sourceconnector.domain;
+package sourceconnector.domain.offset;
 
 public record LocalFileOffsetRecord(
   String key,
