+++ s3-source-connector/src/main/java/sourceconnector/service/batcher/Batchable.java
@@ -1,6 +1,6 @@
 package sourceconnector.service.batcher;
 
-import sourceconnector.domain.MessageBatch;
+import sourceconnector.domain.batch.MessageBatch;
 
 public interface Batchable<T> {
   MessageBatch<T> nextBatch();
